backend/app/plugin/module_gencode/switch_backup/schema.py
# -*- coding: utf-8 -*-

from pydantic import BaseModel, ConfigDict, Field
from fastapi import Query
import datetime
from app.core.validator import DateTimeStr
from app.core.base_schema import BaseSchema, UserBySchema

class SwitchBackupCreateSchema(BaseModel):
    """
    交换机备份管理新增模型
    """
    ip: str = Field(default=..., description='交换机IP')
    name: str = Field(default=..., description='交换机名称')
    backupType: str = Field(default=..., description='备份方式')
    description: str | None = Field(default=None, max_length=255, description='备注/描述')
    status: str = Field(default="0", description='是否启用(0:启用 1:禁用)')


class SwitchBackupUpdateSchema(SwitchBackupCreateSchema):
    """
    交换机备份管理更新模型
    """
    ...


class SwitchBackupOutSchema(SwitchBackupCreateSchema, BaseSchema, UserBySchema):
    """
    交换机备份管理响应模型
    """
    model_config = ConfigDict(from_attributes=True)


class SwitchBackupQueryParam:
    """交换机备份管理查询参数"""

    def __init__(
        self,
        name: str | None = Query(None, description="交换机名称"),
        ip: str | None = Query(None, description="交换机IP"),
        backupType: str | None = Query(None, description="备份方式"),
        status: str | None = Query(None, description="是否启用(0:启用 1:禁用)"),
        created_id: int | None = Query(None, description="创建人ID"),
        updated_id: int | None = Query(None, description="更新人ID"),
        created_time: list[DateTimeStr] | None = Query(None, description="创建时间范围", examples=["2025-01-01 00:00:00", "2025-12-31 23:59:59"]),
        updated_time: list[DateTimeStr] | None = Query(None, description="更新时间范围", examples=["2025-01-01 00:00:00", "2025-12-31 23:59:59"]),
    ) -> None:
        # 精确查询字段
        self.ip = ip
        # 模糊查询字段
        self.name = ("like", name)
        # 精确查询字段
        self.backupType = backupType
        # 精确查询字段
        self.status = status
        # 精确查询字段
        self.created_id = created_id
        # 精确查询字段
        self.updated_id = updated_id
        # 时间范围查询
        if created_time and len(created_time) == 2:
            self.created_time = ("between", (created_time[0], created_time[1]))
        if updated_time and len(updated_time) == 2:
            self.updated_time = ("between", (updated_time[0], updated_time[1]))
