backend/requirements.txt
alembic==1.15.1
APScheduler==3.11.0
fastapi==0.115.2
Jinja2==3.1.6
typer==0.9.0
click==8.1.7
uvicorn==0.30.6
gunicorn==23.0.0
websockets==14.2
httpx==0.28.1
croniter==6.0.0
pandas==2.2.2
openpyxl==3.1.5
SQLAlchemy==2.0.45
pillow==11.0.0
passlib==1.7.4
user-agents==2.2.0
PyJWT==2.9.0
pydantic-settings==2.5.2
psutil==6.1.0
python-multipart==0.0.9
greenlet==3.1.1
bcrypt==4.0.1
itsdangerous==2.2.0
aiofiles==24.1.0
redis==7.1.0
cryptography==45.0.2
rich==13.9.4
pydantic_validation_decorator==0.1.4
loguru==0.7.3
fastapi-limiter==0.1.6
sqlglot[rs]==27.8.0
asyncmy==0.2.9
PyMySQL==1.1.2
asyncpg==0.30.0
psycopg2==2.9.10
aiosqlite==0.17.0
langchain==1.2.0
langchain-openai==1.1.6
langchain-anthropic==1.3.0
langchain-mcp-adapters==0.2.1
# python-socketio==5.14.3
# dmPython

scrapli==2025.1.30
scrapli-community==2025.1.30
asyncping3==5.1.3
asyncssh==2.22.0